Infinite Game Mode (#137)
* Refactored Projectiles

* Added weapon cycling

* Updating Player State for weapons

* Adding Equippables and updating required controls

* Fixing Cycling Bugs

* Toggling between Items

* Finishing up toggling addition

* Added sprites of map and compass to inventory

* Fixing toggling bugs

* Fixed Item Toggling

* Removing multiple firing projectiles

* Fixing Projectile Bugs

* Fixed Player Reset Bug

* Fixing Fire Sprite

* Implementing Link Death Scene

* Fixing some bugs

* Started WallMasterGrabAnimation

* Wallmaster class needs some changes

* Fixing remaing eventManager bugs

* Removing error in constructor calling

* Setting Up

* Setup

* Removing Uneeded Code and making changes suggested

* Removing EventManager and using Commands

* Fixing Errors after Merging

* Fixing Bugs From Merging

* adding wallMasterAnimation

* Small changes to animation of events: LinkDeath, WallMasterGrab, and LevelComplete

* Creating a spawn list for nonCollider spawn

* Adding room generation classes

* Making the entrance room

* Figuring out how to load doors

* Figured out random room loading

* Finalizing doors for random rooms

* Fixing minor bugs

* Adding some new room types

* Initial code for enemy spawning

* Adding stuff to enemy generation

* setting up items

* Adding items and removing some other bugs

* Decreasing health item spawn rates

* Reset candle in each room

* Merging Sprint5

* Fixing Merge Errors

* Fixing errors
