+++ MonoZelda/Dungeons/Parser/Data/IDungeonRoom.cs
@@ -1,4 +1,5 @@
 ﻿using Microsoft.Xna.Framework;
+using MonoZelda.Dungeons.Parser.Data;
 using MonoZelda.Trigger;
 using System.Collections.Generic;
 
@@ -15,6 +16,8 @@ namespace MonoZelda.Dungeons
         List<Rectangle> GetStaticBoundaryColliders();
         List<EnemySpawn> GetEnemySpawns();
         void Remove(EnemySpawn spawn);
+        List<NonColliderSpawn> GetNonColliderSpawns();
+        void Remove(NonColliderSpawn spawn);
         List<ItemSpawn> GetItemSpawns();
         List<TriggerSpawn> GetTriggers();
     }
