+++ MonoZelda/Dungeons/Parser/Data/NonColliderSpawn.cs
@@ -1,0 +1,17 @@
+﻿using Microsoft.Xna.Framework;
+using System;
+
+namespace MonoZelda.Dungeons.Parser.Data
+{
+    [Serializable]
+    public class NonColliderSpawn
+    {
+        public Point Position { get; set; }
+
+        public NonColliderSpawn(Point position)
+        {
+            Position = position;
+        }
+
+    }
+}
