+++ MonoZelda/Dungeons/Parser/ItemCellParser.cs
@@ -1,5 +1,6 @@
 ﻿using Microsoft.Xna.Framework;
 using MonoZelda.Dungeons.Loader;
+using MonoZelda.Dungeons.Parser.Data;
 using MonoZelda.Items;
 using System;
 
@@ -13,7 +14,9 @@ namespace MonoZelda.Dungeons.Parser
             if (Enum.TryParse(cell, out ItemList item))
             {
                 var itemSpawn = new ItemSpawn(position, item);
+                var nonColliderSpawn = new NonColliderSpawn(position);
                 room.AddItemSpawn(itemSpawn);
+                room.AddNonColliderSpawn(nonColliderSpawn);
             }
         }
     }
