--- MonoZelda/Scenes/RoomEnum.cs
@@ -1,9 +1,0 @@
-﻿namespace MonoZelda.Scenes;
-
-public enum Room
-{
-    Room1,
-    Room2,
-    Room3,
-    Room4,
-}
